requirements.txt
python-telegram-bot[job-queue]
python-dotenv
groq
sqlalchemy
alembic
python-dateutil
pytz